firebase-functions/requirements.txt
firebase-admin==6.4.0
firebase-functions==0.1.2
flask==3.0.0
flask-cors==4.0.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-api-python-client==2.117.0
openai==1.12.0
anthropic==0.18.1
pydantic==2.6.1
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
scikit-learn==1.4.0